Simplify asi_nutshell.svg
This was originally created in Google Drawings, now I want to edit it.

This is an LLM-assisted rewrite of the SVG into human-readable code. The
LLM got it 80% right after I just pasted in the old code and said
"simplify this". The rest I was able to tidy up by hand.